# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TeklifPro API",
    "version": "1.0.0",
    "description": "TeklifPro is a Turkish B2B proposal management platform designed to streamline the creation, management, and delivery of business proposals. The API provides comprehensive endpoints for authentication, proposal management, customer management, product catalogs, AI-powered suggestions, and integration with accounting systems via Parasut.",
    "contact": {
      "name": "Sercan Hosgoerns",
      "email": "hosgorens@gmail.com"
    },
    "license": {
      "name": "All Rights Reserved"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.teklifpro.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/v1/webhooks/parasut": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Parasut webhook",
        "description": "Receives webhook events from Parasut. Synchronizes accounting data updates. | Parasut'tan webhook olaylarını alır.",
        "operationId": "parasutWebhook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event": {
                    "type": "string"
                  },
                  "data": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid webhook payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from registration or login"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid input parameters"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
    }
  }
}